LLDP validation: Implement "external" platform port state
Summary:
Series to implement  LLDP validation of received tags against the expected
values, so that faults can be highlighted to the operator ASAP, e.g.  via
the LEDs.

 ---

This part adds the platform code to add a "external" signal to platform port
code.  Initially supporting a "cabling error" external state, as desired by
DC operators while provisioning.  Set a distinctive LED on a port in such a
state.

Cabling error is currently stateless within the port, and so LED will be
cleared by link-state change. This is fine. Essentially the hardware keeps
the state for us as we need it, conveniently.

Other external states may need to add state-tracking to the agent software
side, if or as use-cases come along.

Differential Revision: D13214509

fbshipit-source-id: e21315c212e053cb9affbeac8439cff658750d21