LLDP validation: Apply switch config
Summary:
Series to implement  LLDP validation of received tags against the expected
values, so that faults can be highlighted to the operator ASAP, e.g.  via
the LEDs.

 ---

This part applies the thrift config to the Port.

Note: This would not be needed, if the thrift-generated config code had an
efficient way to lookup ports. The config could be used directly. However,
AIUI, the port list is a std::vector, and so there is no efficient lookup
provided. Hence, the Port class which is indexed by an efficient map, and
which has configuration data copied to it.

If there's an efficient way to lookup a port from LLDP control plane code,
that could be changed.

Differential Revision: D13214507

fbshipit-source-id: afe691428ad8a5d793945d4017f5d0816eeaa2f1