LLDP validation: Thrift switch config for expected LLDP data
Summary:
Series to implement  LLDP validation of received tags against the expected
values, so that faults can be highlighted to the operator ASAP, e.g.  via
the LEDs.

 ---

This part extends the thrift switch config for the port, to add a map of
expected values for each LLDP tag.  Expected values can then be configured
and validated against the received values.

Differential Revision: D13214510

fbshipit-source-id: c9210d84476a7ee9e3b813fc86d9a91ba5408d78