Program queues when enabling ports
Summary:
BcmSwitch::processChangedPorts skips programming ports/queues for disabled
ports with the assumption that the port/queues will be programmed when the
port is enabled (BcmSwitch::processEnabledPorts). However, processEnabledPorts
only enables ports not queues.

This is parcitularly a problem for dataplane BCM tests as
BcmLinkStateDependentTests::applyInitialConfig first applies the config before
bringing up the ports. Minus this patch, it means that the queue configuration
does not get applied for BCM tests.

This does not seem to be a problem for wedge_agent as during agent boot
processChangedPorts is called with ports already enabled which configures
the queue correctly (will investigate why that is the case separately), but
the fix is still valid.

Differential Revision: D13800583

fbshipit-source-id: d1a9c2adf57a280eb70c1974aee8ff2d3fa5cbdc