Add more lldp tests to check lldp and wedge_agent + cli interaction
Summary:
Added the following tests
Start up a emulated wedge agent
- Send 1 lldp packet to it, expect a single lldp nbr to show up
- Down the port over which we sent lldp packet. Expect lldp nbr to disappear
- Bring up the port and send same lldp packet again, expect nbr to show up
  again.

Obviously this does not test the ASIC processing path of LLDP packet. For
that the plan is to test it via a BCM test, by generating the same LLDP
packet and looping it back over a port. ASIC should see the appropriate
packet to show up on the CPU port, and hit the right CPU queue.

Differential Revision: D13830931

fbshipit-source-id: c8629f983587e919af5d7b76a1ba8e1aaa9851f7