Clean up LldpManager::sendLldpOnAllPorts
Summary:
LldpManager::sendLldpOnAllPorts had a flag added to skip
port stat check, just for the purpose of unit tests. This was due
to the fact that the state used in unit tests had all ports disabled.
Instead, its better to just use a state that has ports up and then
exercise the same code path as prod

Reviewed By: SamanKazemkhani

Differential Revision: D13974335

fbshipit-source-id: 560f898d38aaba0bd61a0ea92741cd2a71b19ca1