Unit test to assert that we start sending out lldp packet post configure
Summary:
As titled. As part of incorporating DNE test cases we want
to test LLDP functionality. This tests asserts that immediately post
switch config application, we start sending out packets on (up) ports.

- Other existing unit tests test for various lldp packets getting parsed
correctly
- Fuzz tests assert that we can not crash in face of malicious data thrown
at us
- CLI integration test assert that fboss cli's internal data structures for
representing this info are as expected (presentation is not verified and
depends on the human eye/taste).
- Last piece pending is the ASIC being able to send out and receive LLDP
packets appropriately. We will handle that as part of COPP tests, where
we assert that received packets hit the expected queues.

Differential Revision: D13976400

fbshipit-source-id: 1467234cb4f51bfd46e9e36ea51d49d25b430538