new api for creating SAI entities using SaiAttributeTuple
Summary:
add create2 to SaiApi which expects a given API's type container
to have a nested class called Attributes with a nested class called Create
which implements the SaiAttribute concept and thus has a SaiAttrV for use
in calling the underlying sai create api. PortApi is the first concrete
example to demonstrate the technique.

Changing all of them would be a huge, confusing diff, so I think it is
very worth it to add create2 temporarily while we move away from create.
In the end, the intent is to get rid of the current create and replace it
with create2 (i.e., in the end, create2 will be renamed to create)

Differential Revision: D13453955

fbshipit-source-id: d210ea6342569ce40d809a14fdf5ee56c7cc6fbf