Unittest for when h/w and s/w switches get out of sync
Summary:
In case of table overflow (CAM, FIB) overflow
we continue running the wedge_agent with the current set
of routes, but mark the h/w & s/w switches as being out of
sync. Once this happens the SwSwitch tries to get them
in sync with every new update (in the hopes that the
new update freed up some space). We are going to
test the h/w (ASIC) overflow in a BCM test. Test the
SwSwitch logic in unit test.

Reviewed By: boryas

Differential Revision: D14032693

fbshipit-source-id: 10b5cff250d34de5c5e677e5b5873debab353176