RefMap for ref-counting in SaiSwitch
Summary:
There is a pattern in SaiSwitch (and BcmSwitch for that matter) where
we have shared ownership of some object, which we want to be ref-counted, but
need to be able to do lookup for. The simplest example is NextHopGroup
(i.e., ECMP group). Any route can be the first route to create an ecmp group
for a particular set of next hops, and we should keep it around as long as
any route (possibly not the original one) still refers to it.

To that end, create a map class which stores weak_ptrs to the objects, and
hands out shared_ptrs to the creator/users. The shared_ptr is imbued with a
lambda which will delete the empty entry from the map when the ref count
goes to 0.

I think that it is likely we will have to add the ability to iterate over
the map, but I will defer that until we actually need it.

Differential Revision: D13454790

fbshipit-source-id: 0125ea404fbbccf9ed146151effe9fd289cfa80b