Remove cfg::HashingAlgorithm::CRC16_XOR8
Summary:
In flow-based RTAG7 output selection, a flow identifier is calculated
by a user-configurable hash function. Broadcom recommends applications use an
XOR-folded hash for this purpose and a strict CRC function for the RTAG7 hash
calculations.

To enforce this guideline, only CRC hash functions are exposed in
cfg::HashingAlgorithm and the flow identifier hash function is hard-coded to
CRC16_XOR8.

Differential Revision: D14126583

fbshipit-source-id: 609391e7d3f127b602ec233e4804225d7041c2f1