add SaiVirtualRouterManager
Summary:
SaiVirtualRouterManager will be used to manage VRFs. In practice, we
will always have just one global VRF. In fact, we expect the SAI adapter to
provide a default VRF whose id is exposed by a switch api attribute. For this
reason, the currently supported model is to create an un-managed default
SaiVirtualRouter at the creation of the manager, and to store it with under
RouterID(0). This makes adding support for creating VRs, and for handling
a missing default VR as flexible as possible. Other SaiManager code using
VR looks the same as any other manager code, doing lookups based on SwSwitch
IDs and getting back a pointer to an object with sai ids/attributes.

Differential Revision: D13582396

fbshipit-source-id: 849f103934da0af2a01c9740a7cf58921fb94fd1