add periodically flapping lpmode
Summary:
This adds logic that will clear and then reset the power
class of a transceiver periodically. We are seeing some transceivers
on minipack get in to a weird state where the only way to recover is
to reset or to flap lpmode. This would recover those ports after some
interval.

Differential Revision: D14174061

fbshipit-source-id: f22f14021fb32798d85258ca015e0a2e5a7ac0a4