Don't skip QoS maps for default queue
Summary:
We must explicitly configure QoS map for default queue (queue 0) or else,
the packets will be processed by default queue as expected, but the dscp
marking will be removed (dscp set to 0). Note that this behavior is
different from ACLs. CS7475492 raises this question with Broadcom, but in the
interim, unlike previously, we don't skip creating QoS map for queue0.

Here is a post giving additional context:

https://fb.workplace.com/groups/668135506568401/permalink/2330088780373057/

Additional fixes that had to be part of this patch (or else tests will fail):
- Agent now has explicit check to verify if the supplied QoSPolicy in the
  config has mappings for ALL dscp values, or else it throws error.
- Existing QoSPolicy state tests as well as BCM tests were not configuring
  all mappings. The helper functions used by these tests is modified to
  implicitly configure missing mappings as mappings to default queue queue0.

Differential Revision: D14089797

fbshipit-source-id: ed754de91a0202340e7eb8175a2bb088b523b447