Publish hw_out_of_sync values always and make counter a bool (in sync or not)
Summary:
We had used a TLCounter for hw_out_of_sync (we could not apply
the desired updates to the switch due to FIB, host table overflow).
Then we would inc/dec this counter as we got state update failures.
However, that meant that the value of the counter was not set till
after the first failed update and thus did not get published till then.
Since we just want to track a binary state in sync vs not, jut directly
publish the counter.
Finally, one perceived advantage of TLCounter could be that it could
keep track of how many attempt for state update application failed. This
perception is wrong, as
- We requeue the failed update to be applied with the next state update (
in the hope that if the next update frees up some table space then this
updated may suceed). So in case of persistent failure the counter will
keep getting incremented for the same failure.
- On returning back to h/w s/w sync (space was freed in the table and
we were able to get h/w and s/w states in sync), we were decrementing the
counter by 1. However considering the comment above that would not be
sufficient to get the counter back to 0, which was the intention here

Reviewed By: phshaikh

Differential Revision: D14205469

fbshipit-source-id: c431b07be844a7d738b92e77a0b36db3dea3992a