Add a unit test for recieving unsolicited NDP adv
Summary:
In D14154090, I noticed that receiving a unsolicited
NDP adv crashed fake_wedge_agent. My first thought was that this
was triggering a bug in our packet handling path. But on further
investigation (as a follow up to adding this unit test),
it turns out that the bug is actually in FakeSdk. And we crash
when sending out a ndp neighbor solicitation in response to this
packet. FakeSdk fix is coming in another diff. However, this test
was atleast useful in asserting that agent code does handle
unsolicited nbr adv correctly. So let it stand.

Differential Revision: D14200148

fbshipit-source-id: 4be0542642b2edb3f2b7d48cbeb0a40065b2a5ce