Set DST_DISCARD on routes with a DROP action.
Summary:
Till now we we were only setting DST_DISCARD flag
on the drop egress. Unfortunately this meant that configuring
a cstat to catch DST_DISCARD did not actually catch packets
hitting these routes (i.e. setting it just on egress was not
transitive to the route pointing to this egress w.r.t to this
counter). This is what tripped me when playing with this last
time.

So update route programming to set this flag for
routes pointing to drop egress.

TODO: Do this for host entries/host routes as well and bring
back pending entries pointing to DROP

Reviewed By: boryas

Differential Revision: D14267712

fbshipit-source-id: bff826389a39e429a00375478b159375f86c36c0