parameterize RefMap by map container to allow for flat_map
Summary: We need to use a SwitchState NextHopSet as the key in a RefMap for next hop groups. NextHopSet doesn't support hashing, so we can use a flat map as in BcmSwitch. However, that means RefMap must be parameterized by the map type rather than just blindly using unordered_map. The different customization points of various map templates means we need to specialize them down to just <K, V> before use in RefMap.

Differential Revision: D13604054

fbshipit-source-id: 641e3d5d9c484e81227c5dbe3889ef3c8af3db3c