future::onError -> future::thenError
Summary:
onError famously doesn't have proper support for executing the
callback on a supplied executor (regardless of via or anything else).
It appears that thenError is the new API, and we are already seeing some
build failures in oss builds on diff urging us to move. We don't have
too many use cases so it seemed a reasonable switch to make.

Differential Revision: D14327215

fbshipit-source-id: 47b5896345a17c9aea4162ae9db91b86f74ea05b