RTAG7 output selection
Summary:
Debugging efforts over the last several weeks revealed that the process
by which output is selected from the RTAG7 engine was misunderstood. A pipeline
(logically) independent of the RTAG7 engine is responsible for assigning a
load-balancing application (eg., ECMP, link aggregation, link aggregation group
failover) output from the RTAG7 engine.

This change rectifies several ways in which this pipeline was misprogrammed or
under-programmed. They include:

- RTAG7 output selection for link aggregation has been changed to flow-based
  from port-based (which it had been silently defaulted to)
- Macro-flow ID computation is now _explicitly_ configured (see
  BcmRtag7Module::programMacroFlowIDSelection())
- The RTAG Flow-Based Hash Table is now _explicitly_ configured (see
  BcmRtag7Module::programFlowBasedHashTable())

The RTAG7 output selection pipeline is programmed in a data-directed manner by
the OutputControl structure.

Differential Revision: D14253301

fbshipit-source-id: 4802b91f51610192f1e4d2e4899131d1dd7deeca