fix state reversal on overflow
Summary:
During reliability week, I wrote tests for resource exhaustion. Purpose of these tests was to overflow BCM HW tables, especially host and L3 defip tables. However as I was testing resource exhaustion, I found that BcmSwitch was crashing while handling overflow. (P60327987)

Investigation revealed that node was being removed from flat map while iterating flat map, and iterator was not aware of this deletion. This is incorrect interaction of iterator and deletion of node. It is possible to erase map entries while using iterator, but that means iterator must be reset, whenever any node in map is removed. Current implementation, did not do that.

To fix this, I am proposing an alternate implementation. In this implementation, a map entry is not removed while loop is active. Instead "Deltas" (and any other information required to restore state such as router id) is collected on overflow. These deltas are then discarded by reusing existing mechanism of reversing the state. This is simplest fix to overcome the problem, alternate solution would involve implementing a loop with erase that will allow resetting an iterator. This may also need changes in parts of code, which otherwise are working and are stable.

Differential Revision: D13080582

fbshipit-source-id: 0ed48ecf3a2795ef30820702ffc5d2b4595fe741