Retire no longer used optimization
Summary:
Previously we used to have a optimization where
when *all* bcm_tests were run a single process, we did not
recreate and attach to BcmUnit every time. This was deprecated in
D7888008 and D13595532 due to it not being truthful in reproing
all SDK bugs. So now kill the code used to accomplish this.

A nice side effect of this is that it completely unifies the
way in which wedge_agent and bcm_tests create and init BcmSwtich

Reviewed By: shrikhare

Differential Revision: D14319748

fbshipit-source-id: 7ab5e9bb4eee58cf2fb8c52cf6a0c8a4169ce47b