Make bootType available at the HwSwitch level
Summary:
Concept of cold vs warm boot applies to all HwSwitches.
I am trying to abstract out a HwTest from BcmTest in the hopes
that we can use some of this for SaiTests as well.
For now I have still retained the interface only nature of HwSwitch,
but this does come at the cost of some code dup. If this gets
cumbersome, we should consider putting members and implementation
in HwSwitch

Reviewed By: shrikhare

Differential Revision: D14305206

fbshipit-source-id: 5b7fc2f7cfe7a6ceaac1394cd92252ca21d1eb4a