SaiNextHopGroupManager
Summary:
This class manages next hop groups. The lifetime of SaiNextHopGroup is determined by the SaiRoutes which refer to it, so SaiNextHopGroups are stored in a RefMap here.

When a neighbor referenced by one of the next hops in this next hop group becomes resolved or unresolved, we need to modify the next hop group appropriately, which is the reason for the map from neighbor entry to next hop group. Currently, this map is buggy as it is not incorporated into the RefMap delete code. This can probably be fixed with a proper dtor for SaiNextHopGroup, or by making a fancier "indexed" map.

Differential Revision: D13604052

fbshipit-source-id: 85b92214673fa06dbebc6be4429be65d7310168c