Helper function to compute derived h/w counters
Summary:
For in_non_pause_discards we compute them by taking
in_discards and subtracting pause frames from them. There are
a few complications that make this computation slightly convoluted.
- Counters are of different width - in_discards is 31 bit and pause frame
counter is 39 bits.
- Counters can rollover at different times
- We can only begin our computation after the second collection interval
since we need previous values to compute the offset.

So idea here is to abstract this logic into a helper function that. We
will reuse this for computing in discards sans pause frames and dst blackhole
discards. Logic is
- If any of the prev counters is uninit, return 0 increment (can't compute
increment),
- If any of the cur counters is < prev value, return 0. Counter rolled over and
we skip computing increment. This could be made smarter by encoding bit lengths
into computation above. But it seems like a overkill as opposed to just
skipping that single sample.

Reviewed By: boryas

Differential Revision: D14297628

fbshipit-source-id: d98b4acdaa20dd774425568c588b935132528458