Enacapsulate calculations in CounterCurAndPref
Summary:
Mainly incorporate boryas's suggestions from D14297628.
One extra addition is to play defensive and not crash the
application in case current is uninitialized. Instead log
a error message. There is a more general case of counter
value being -ve. Which we never expect outside of the first
collection interval, where prev maybe uninit (-1). So don't
add any code for that, rather let that show up as bogus increments
and log noise. If we our assumptions are wrong, we will see
log noise and -ve counter values for raw counters in ODS.

Reviewed By: boryas

Differential Revision: D14418356

fbshipit-source-id: 4868410e06e42b56dc3f0b050ec6286fa2054a36