give transceivers time to handle low power set
Summary:
Looks like flapping lpmode in short succession does not
always stick with certain transceivers. This adds a sleep between the
two operations.

I will also look in to extending the logic to only force reset like
this when the port has been down for an extended period of time.

Differential Revision: D14442988

fbshipit-source-id: 8d0be266c9ced76f71d331426308145d3a7b3c2a