add support to generate v6 addresses and prefixes
Summary:
Extending IPAddress & prefix generation to v6.

For v6, uint32_t is not enough "id space". Accordingly, introduced IdV6 which is pair of 64 unsigned ints.

Consider (A,B) as ID for v6 ID-space, where A & B both are unsigned 64 bit integers. A is most significant 64 bits, while B is least significant 64 bits.

For prefix generation, (A,B) to be right shifted by "mask". This means both A & B must shift to right by "mask" bits, with caveat. Most significant "mask" bits of B, occupy least significant bits of A.

This logic forms the heart of this diff.

Differential Revision: D14294252

fbshipit-source-id: 143f9a8fb463ada3d358e0c9d022121a85bfde0e