Extract out a ConfigFactory abstractions for constructing configs for bcm_tests
Summary:
Extract out a ConfigFactory abstractions for constructing configs for
bcm_tests. I moved this out of the kitchen sink BcmTestUtils.* and moved
it up to non facebook dir. Immediate motivation is bcm_test-oss. But once
that hurdle is crossed we can pull it even higher. Right now the config
generation uses a single bcm api (to get max port speed). We can simply
parmeterize config generation with it.

Reviewed By: shrikhare

Differential Revision: D14432970

fbshipit-source-id: 36dce0f6418d4ef0f273cbbee68bab12abbdcccf