Test to assert behavior of inPause frames w.r.t. inDiscards
Summary:
BRCM ASIC bumps up the inDiscard counter on receiving
a inPause frame regardless of rx pause being enabled or not.
This behavior depends on MMU being configured as lossless vs
lossy. Since we always setup MMU as lossy, both test cases
assert for similar counter increments.

What is surprising is that even though BRCM counts received pause
as a in_discard, it **does** act on it in terms of pausing traffic.

Reviewed By: boryas

Differential Revision: D14298298

fbshipit-source-id: dccaf0f8ba71217246f937b46ae470f1bffa9b9e