SaiRouteManager
Summary:
This diff introduces a route manager which handles route deltas from SwSwitch.

The primary difficulty in this implementation is handling unresolved l2 next hops for forwarding routes. The logic is currently as follows: If we intend to trap or drop, just trap or drop, don't set the next hop attribute. Otherwise, we mean to forward. We createOrRef the next hop group that this route wants to forward over. If that next hop group is empty, we set the action to trap anyway. Otherwise, we set the next hop id to that of the next hop group. Going forward, if a next hop group ever becomes empty, or stops being empty, we lookup the routes using that next hop group and set them trap or the next hop group as appropriate.

Differential Revision: D13604042

fbshipit-source-id: 583c7f6ff7d67943ab0ad504d87248cadd17eabc