Enable Travis Build Bot for FBOSS's open source builds
Summary:
When we do a commit, travis-ci.org (https://travis-ci.org/facebook/fboss)
will automatically try to build the FBOSS code.  This diff tries to
enable the FB 'travis build bot' to add comments to the FB internal diff
for whether the build passed or not.

Unfortunately, this information is only posted *after* the diff lands and
to the closed diff, but it's better than nothing <shrug>

Not sure if this change to the notifications is sufficient but from
RTFS'ing around, it seems to be necessary.

Reviewed By: simpkins

Differential Revision: D14536022

fbshipit-source-id: 6ee0be563fccb280484bb4db638ff8e737fa92c7