SaiPlatform and SaiPlatformPort
Summary:
In order to create a full wedge_agent binary that uses SaiSwitch, we need to create a Platform object.

Unfortunately, the existing hierarchy of Platform implementations are fairly tightly coupled with BcmSwitch, so they are not readily reusable. For this reason, we will create a new implementation of Platform and PlatformPort, and try our best to lift common code from our existing Platforms.

Differential Revision: D13604046

fbshipit-source-id: eaa938d7e0e8443b4d7be81fb233432f4a6feacd