Add first OSS bcm_test
Summary:
As title. Note that while this runs fine with FB build of bcm_test,
bcm_test-oss still croaks while initing the device. So the success here is
only in being able to build this test with opennsl, and not being actually
able to run it on a FBOSS switch. This is the same place we are with
wedge_agent-oss currently (builds but unable to run on a in house switch).

Working with BRCM to get over the device init problems

Reviewed By: shrikhare

Differential Revision: D14450589

fbshipit-source-id: df7a7bc604631c40d9c4151644bbb32e9fb320c8