lazy-load massive oui_vendor map
Summary:
Since we do this at import time, it can have an impact on
startup time for our cli.

Our cli is abysmally slow right now and just invoking -h takes 6+
seconds on some rsw devices I was testing on.

With this change, it shaves off about 0.3-0.4s. Not a ton, but
probably worth it.

Reviewed By: shrikhare

Differential Revision: D14509307

fbshipit-source-id: f237f36cb120a3ba7dd259d682b4fb0e347605c5