Fix sflow sampling and write bcm_tests to verify functionality
Summary:
As titled. Sflow sampling seems to have been broken
during 6.5.13 migration work. Fix it now and write tests for
verification.
Note that I got rid of the call to set sampling destination.
Right now we always use CPU as the destination, which is the
default anyways. TH onwards, we can set destination to be a
mirror as well. Since we don't yet expose that functionality
get rid of this API call completely. We will add it when
we support such functionality

Reviewed By: phshaikh

Differential Revision: D14631198

fbshipit-source-id: 27fb94805d63e2cca78123a9eaaef6dc5dcaf42d