Fix a subtle race in sendPktSync code
Summary:
Our approach to syncPacketSend was thus
- In packet send path
 Acquire lock
  - set syncPacketSent to false
  - sendPacket
  - Wait on a CV for syncPacketSent to become true
- In packet callback
 - Acquire lock
  - set syncPacketSent to true
  - Signal CV to unblock syncPacketSent

This assumed that packet callback always happened in a separate thread. However
this is not true. I added some hacky logging to this code - P61763216
And here are the deadlocked and non deadlocked examples - P61763202 (deadlocked
is first). Notice how in the case of deadlock the thread id of packet send
code and callback are the same. Hence if we send packet while holding the lock,
if the callback happens in the same thread, the callback will want the lock
which is already being held. Hence deadlock. There are 2 solutions I could
think of - recursive mutex or give up the lock while sending the packet.
The second is more appealing and that's whats implemented here. The
sendPacketImpl does not update any data structures that need to be
protected under a lock.

Reviewed By: shrikhare

Differential Revision: D14585964

fbshipit-source-id: f00458b4b3b0ea3e569c495c0bb21d501fe43886