fix failing bcm test
Summary:
Landing D14484357 led to failures in some BCM tests (details in task).

In D14484357, equivalence was established by comparing two egress structures. This was done using comparator operator ==, in addition to intf, port, mac and label, it also compares "flags" which is failing this test.

Restoring the previous comparison for quick fix.

For longer term, need to see as to why flags differ and if comparing them should really be part of comparison operator

Reviewed By: shrikhare

Differential Revision: D14671522

fbshipit-source-id: 4ad70a349c63f60fc7e3399dfa006e21f4ebfaad