fix segfault in getRouteTable and getRouteTableByClient
Summary:
Without a non-reference handle to SwitchState, updates happening
in parallel to the execution of the for loop in this api call can result
in freeing the memory held by a shared_ptr that we only have a reference
to. As a result, our shared_ptr has inconsistent state and we can do a
double free or dereference freed memory. This diff fixes the issue,
though I suspect we can do more, like banning the use of
`const std::shared_ptr<Node>&` return values. I will do a bit more research
along those lines to see if there are some easy ways to make the SwitchState
api safer to use.

Differential Revision: D14689962

fbshipit-source-id: 796fd8292d028c6f1a19c4e5d1a1b0efcd71c583