convenient wrapper to access host entries
Summary:
incRefOrCreate and deref methods are used to create and/or access host/ecmp host entry and relinquish reference to it, I found it somewhat error prone and cumbersome to use.

Its possible to simply get hold of BCM host or BCM ECMP host and forget to "deref" it.

Also its somewhat inconvenient to know "key" construction to access host entry.

Some forwarding application (like label switching) are agnostic to multi-path or non-multipath egress (or host) object, and application only care if egress is resolved or not (programmed or not).

Providing RAII style convenient wrapper BcmHostReference to access host table entry, if entry is already fetched, then it is not fetched again. To update this entry, discard (or reset) a reference and get new one.

Also hiding details of key construction inside this class, which otherwise would get repeated at many places.

Reviewed By: boryas

Differential Revision: D14524582

fbshipit-source-id: ee83c7ec1ef44ecd3ec6cafc7b61f84082e737aa