label switch action
Summary:
label switch action, encapsulated in an object.

it obtains and holds a reference to BcmEcmpHost or BcmHost  depending on if label switch action is multi path or not.

SDK does not provide any API to "update" entry in place, accordingly constructor and destructor are the only relevant methods for this object.

Not supporting label push in this diff, as it requires creation of MPLS tunnels - this will happen in subsequent diff.

Reviewed By: boryas

Differential Revision: D14401419

fbshipit-source-id: 13c80dbe72398f8c43d8deedea27f5b3329f5dd6