handle interface subnet routes
Summary:
Each subnet on an interface needs two routes:
* one to the exact ip address itself to punt traffic to FBOSS. This route does not
come through the FIB from SwSwitch, so we will handle it separately.
* one to the subnet programmed to trap traffic to the subnet to trigger neighbor resolution.

in SAI, the second type of route is programmed by making a rif_id
the next hop id attribute of a route object. We detect if a route
from the SwSwitch is CONNECTED and if so, we use the appropriate rif_id
as the next hop.

Differential Revision: D14716185

fbshipit-source-id: b1a4356b1c80a9ce5fadfc04befe38e29b330744