bcm label tunneled egress
Summary:
A further specialization of "labeled egress".

In addition to being "labeled egress", this is associated with specific labeled tunnel.

Tunneled Egress is bound to a labeled tunnel - a labeled tunnel is created (if none exists) - while constructing tunneled egress.

This egress object represents "FEC" which will be imposed by remote label distribution peers.

Note that a tunnel could be shared by many egresses if these egresses are over same interface to same LDP peer but with different tunnel labels

Differential Revision: D14656508

fbshipit-source-id: 41861a599cd8f5afb2a690c604221d537f52502b