using bcm host references in bcm routes
Summary:
A major change in BcmRoute.

A change is major not because of its size but because it touches a core area of route programming in BcmSwitch.

Instead of  incrementing & decrementing BcmHost or BcmEcmpHost reference counts and maintaining scope guards to do proper cleanup, using BcmHostReference (D14524582) .

BcmHostReference is further extended to return reference to host based on key.

BcmRoute now maintains nexthop reference and host route reference.

Note -
BcmHostReference  makes it transparent whether underlying route is IP2IP or IP2MPLS

Reviewed By: boryas

Differential Revision: D14688280

fbshipit-source-id: b782574c7ee15e8fded22833bc9a426254f68f32