reset reference to host route prefix before programming
Summary:
reseting host route host reference before "programHostRoute" call.

if next hops of host route change, then BcmEcmpEgress for the route must be updated.

Accordingly, setEgressId is invoked on Host Route host reference in "programHostRoute". This decrements a reference of ECMP egress by 1,

If "host reference to host route" is not reset then getBcmHost(BcmHostKey(vrf, hostRoutePrefix)) in "programHostRoute"  return  "stale" BcmHost entry.

This entry is already "addedInHw_" and so egress to host route is never updated. In ASIC host route continues to refer old ECMP egress.

In BcmRoute's program, when host reference is updated with new host reference (created in programHostRoute), ref count of ECMP egress reaches 0, and is thus "destroyed" BUT it still is in use in ASIC which leads to E_BUSY error and subsequent crash,

Detailed log analysis @ T43048093

Reviewed By: boryas

Differential Revision: D14929389

fbshipit-source-id: 1b4d1be607b21fe9010dfd342d7a50987ab28d39