Initial gearbox stats
Summary:
This adds fb303 stats for gearboxes. We are collecting two stats:
 - fec uncorrectable errors (fboss.agent.eth2/1/1.xphy.system.fec_uncorrectable)
 - signal to noise ratio (fboss.agent.eth2/1/1.xphy.system.lane2.snr)

The first counter is one per port+side. The second is for every lane
used on either side of the gearbox for a given port.

We may decide that snr is not worth collecting on every port. We can
do this both via filtering in our monitoring config or skipping
populating in to fb303 even to save some mdio cycles.

Reviewed By: boryas

Differential Revision: D14142504

fbshipit-source-id: 147d58da783b175a7339818ee689ea512f17de96