reprogram labeled egresses on neighbor changes
Summary:
Reprogram labeled hosts (host entries with labeled egresses) in response to neighbor entry changes.

This is done transparently to BcmSwitch and BcmHostTable program all hosts (labeled or unlabaled) to a a port / trunk / cpu in response to neighbor discovery changes on an interface.

if neighbor is discovered, an unlabeled host and labeled hosts (there could be many) get programmed to port/trunk.

if neighbor entry expires, an unlabaled host and labeled hosts get programmed to CPU (and/or drop for labeled hosts).

if neighbor entry changes, an unlabaled host and labeled hosts get reprogrammed.

complexity of reprogramming all labeled hosts is logarithmic - O(logN) where N is number of labeled host entries. this is by virtue of two choices in code
1) labeled hosts are maintained in flat_map which is ordered random access iterator
2) equal_range which facilitates identifying labeled hosts based on BcmHostKey, has logarithmic complexity.

Reviewed By: boryas

Differential Revision: D14712056

fbshipit-source-id: 2f0e0c0d9a545657354299106c13535315c9450b