factoring processing of neighbor entry delta
Summary:
Major surgery of code of how neighbor entry delta is handled.

In response to neighbor entry changes, hosts are either pointed to CPU (for entry deletion) or pointed to port or trunk (for entry addition or change).

This logic remains the same, however programming of "host" to CPU/port/trunk is moved in BcmHostTable.

This is done on-purpose. Below is justification for refactoring -

For MPLS, a neighbor entry changes should trigger updating "labeled" hosts as well.

For neighbor entry addition/change, all hosts (referring to that neighbor) but using labeled egresses will be programmed to port / trunk.

For neighbor entry deletion, all hosts (referring to that neighbor) but using labeled egresses, will be programmed to "drop" (TBD between to CPU or to drop).

Current code in BcmSwitch to handle host programming in response to neighbor entry changes appeared inflexible enough that motivated me to do this refactoring.

Keeping logic (and design) simpler, less complicated and to avoid cluttering code in BcmSwitch  to deal with "different kinds of hosts" on neighbor changes, decided to move the logic to handle host programming in BcmHostTable.

Accordingly following methods are introduced
- programHostsToTrunk
- programHostsToPort
- programHostsToCPU

Further changed return type of registerNeighbor and unregisterNeighbor, they're not required to return anything as BcmSwitch will program neighbors through host table.

Subsequent diff will have logic to program labeled hosts along with "unlabaled" host.

Reviewed By: boryas

Differential Revision: D14707950

fbshipit-source-id: 9ee6b616ada5cdd43e6eb345497959df1202b82d