Assert that all ports are disabled on cold boot
Summary:
As titled. We are looking to enable packet handling
post init and not have to wait till configured state. On warm
boot this works fine, but for cold boot it means we would be
handling packets before config is applied and thus using cold
boot state to do this. This is only a concern for packets coming
from host as the front panel ports are disabled in cold boot state.
Assert that in a bcm test

Reviewed By: saifhhasan

Differential Revision: D15006884

fbshipit-source-id: 2c78f39bf7caafe2ce0f23e151b7b57d75add578