Track FIB size statistics
Summary:
The routes.v4 and routes.v6 counters report the total number of routes
stored in RouteTableRib's (across all VRFs).

If the standalone RIB is enabled, these counters instead report the total number
of routes stored in ForwardingInformationBase's (also across all VRFs).

Note that the semantics of these counters are changed across the transition to
the standalone RIB. More specifically, under the standalone RIB these counters
report the total number of routes in the FIB. Previously, these counters
reported the total number of routes in the RIB.

This change in semantics is justified by the RIB size being tracked elsewhere.

Differential Revision: D13686774

fbshipit-source-id: ccc4264d32e64274ebbf6671ce02e0890efaccd9