Fix neighbor entry serialization
Summary:
Serialize/deserialize neighbor entry state as well.
This did not matter much till now, since we were reconstructing
neighbor entries from warm boot cache rather then serialized
switch state https://fburl.com/md3kvy4c
In D15010203, I changed that to go entirely to using deserialized
switch state. On doing so neighbor bcm_tests started failing. Reason
for this was
- To CPU/pending entries were not skipped over when recovering from
warm boot cache. And the tests just looked at their state in h/w.
This is just lucky, in the wedge_agent, FIB SYNC would come and
blow away such entries from HW if there was another probe for
this neighbor address.
- With D15010203, recovering and replaying the entire switch state
we transitioned these entries to unverified, which caused tests to fail.

In the new approach, just serialize and deserialize as is. The motivation
for going to UNVERIFIED for entries which have a port, mac seems to be
to immediately reverify then after a warm boot. But this does not really
seem critical. If the entry trully became unreachable after warm boot
(neighbor disappeared or changed MAC), it will be be expired in due time.
Given the number of probes we do (300) this doesn't seem like a worth
the inconsistency it adds to our code

Reviewed By: saifhhasan

Differential Revision: D15042056

fbshipit-source-id: 2066e4f54717b1ee2adbfacd23862eb7c889ab80